refactor(site/src/pages/AgentsPage/components/ChatConversation): extract ReasoningDisclosure into its own file
Moves the inline ReasoningDisclosure component out of
ConversationTimeline.tsx into a dedicated file so the upcoming
collapsed-by-default redesign can be iterated on (and story-tested)
in isolation. No behavior change.
